Max2babylon:  - Light, mesh, camera animations Babylon.js:  - Support for lights animation  - Fixing issue with missing quaternion.subtract function

+++ Exporters/3ds Max/BabylonExport.Entities/BabylonAnimationKey.cs
@@ -6,7 +6,7 @@ namespace BabylonExport.Entities
     public class BabylonAnimationKey
     {
         [DataMember]
-        public float frame { get; set; }
+        public int frame { get; set; }
 
         [DataMember]
         public float[] values { get; set; }
